Transition from glicko2 to new rating system
New minimum value is 80, if rd is smaller than 80, it means the player has not yet played a game with the new rating system. So I transition him by multiplying by 80/60.25, 60.25 is the old minimum value. This is hacky, just do it so we can roll it out in S5 and test before S6 begins.